+++ .gitignore
@@ -1,2 +1,5 @@
-/vendor/
+.idea/runConfigurations/
 .idea/workspace.xml
+/vendor/
+build.php
+out/
