+++ src/pathfinder/Pathfinder.php
@@ -1,9 +1,6 @@
 <?php
-
 declare(strict_types=1);
-
 namespace pathfinder;
-
 use pathfinder\command\PathfinderCommand;
 use pathfinder\entity\TestEntity;
 use pocketmine\data\bedrock\EntityLegacyIds;
@@ -14,16 +11,15 @@ use pocketmine\plugin\PluginBase;
 use pocketmine\Server;
 use pocketmine\world\World;
 
-class Pathfinder extends PluginBase {
-    public static Pathfinder $instance;
 
-    protected function onEnable(): void{
-        self::$instance = $this;
+class Pathfinder extends PluginBase{
+	public static Pathfinder $instance;
 
-        Server::getInstance()->getCommandMap()->register("pathfinder", new PathfinderCommand());
-
-        EntityFactory::getInstance()->register(TestEntity::class, function(World $world, CompoundTag $nbt) : TestEntity{
-            return new TestEntity(EntityDataHelper::parseLocation($nbt, $world), $nbt);
-        }, ["TestEntity"], EntityLegacyIds::VILLAGER);
-    }
+	protected function onEnable(): void{
+		self::$instance = $this;
+		Server::getInstance()->getCommandMap()->register("pathfinder", new PathfinderCommand());
+		EntityFactory::getInstance()->register(TestEntity::class, function (World $world, CompoundTag $nbt): TestEntity{
+			return new TestEntity(EntityDataHelper::parseLocation($nbt, $world), $nbt);
+		}, ["TestEntity"], EntityLegacyIds::VILLAGER);
+	}
 }
