+++ src/pathfinder/algorithm/Algorithm.php
@@ -1,126 +1,121 @@
 <?php
-
 declare(strict_types=1);
-
 namespace pathfinder\algorithm;
-
 use Closure;
+use pathfinder\algorithm\path\PathResult;
 use pathfinder\Pathfinder;
-use pathfinder\pathresult\PathResult;
 use pocketmine\math\AxisAlignedBB;
 use pocketmine\math\Vector3;
 use pocketmine\scheduler\Task;
 use pocketmine\scheduler\TaskHandler;
 use pocketmine\world\World;
+
 use function microtime;
 use function round;
 
-abstract class Algorithm {
-    protected AlgorithmSettings $settings;
-    protected AxisAlignedBB $axisAlignedBB;
 
-    protected ?PathResult $pathResult = null;
-    protected ?Closure $onCompletion = null;
+abstract class Algorithm{
+	protected AlgorithmSettings $settings;
+	protected AxisAlignedBB $axisAlignedBB;
+	protected ?PathResult $pathResult = null;
+	protected ?Closure $onCompletion = null;
+	protected int $tick = 0;
+	protected float $startTime;
+	protected float $startTickTime;
+	protected bool $running = false;
 
-    protected int $tick = 0;
+	public function __construct(protected World $world, protected Vector3 $startVector3, protected Vector3 $targetVector3, ?AxisAlignedBB $axisAlignedBB = null, ?AlgorithmSettings $settings = null){
+		$this->axisAlignedBB = $axisAlignedBB ?? AxisAlignedBB::one();
+		$this->settings = $settings ?? new AlgorithmSettings();
+	}
 
-    protected float $startTime;
-    protected float $startTickTime;
+	public function getWorld(): World{
+		return $this->world;
+	}
 
-    protected bool $running = false;
+	public function getStartVector3(): Vector3{
+		return $this->startVector3;
+	}
 
-    public function __construct(protected World $world, protected Vector3 $startVector3, protected Vector3 $targetVector3, ?AxisAlignedBB $axisAlignedBB = null, ?AlgorithmSettings $settings = null){
-        $this->axisAlignedBB = $axisAlignedBB ?? AxisAlignedBB::one();
-        $this->settings = $settings ?? new AlgorithmSettings();
-    }
+	public function getTargetVector3(): Vector3{
+		return $this->targetVector3;
+	}
 
-    public function getWorld(): World{
-        return $this->world;
-    }
+	public function getAxisAlignedBB(): AxisAlignedBB{
+		return $this->axisAlignedBB;
+	}
 
-    public function getStartVector3(): Vector3{
-        return $this->startVector3;
-    }
+	public function getSettings(): AlgorithmSettings{
+		return $this->settings;
+	}
 
-    public function getTargetVector3(): Vector3{
-        return $this->targetVector3;
-    }
+	public function getPathResult(): ?PathResult{
+		return $this->pathResult;
+	}
 
-    public function getAxisAlignedBB(): AxisAlignedBB{
-        return $this->axisAlignedBB;
-    }
+	/**
+	 * $closure(?PathResult $pathResult) is called, when the path has been found or the pathfinder has a timeout
+	 *  => $pathResult is null, when no path was found
+	 */
+	public function then(Closure $closure): self{
+		$this->onCompletion = $closure;
+		return $this;
+	}
 
-    public function getSettings(): AlgorithmSettings{
-        return $this->settings;
-    }
+	public function isRunning(): bool{
+		return $this->running;
+	}
 
-    public function getPathResult(): ?PathResult{
-        return $this->pathResult;
-    }
+	public function start(): self{
+		if ($this->running)
+			return $this;
+		$this->running = true;
+		$this->startTime = microtime(true);
+		$this->tick = $this->settings->getMaxTicks();
+		if ($this->tick > 0) {
+			Pathfinder::$instance->getScheduler()->scheduleRepeatingTask(new class($this) extends Task{
+					private Algorithm $algorithm;
 
-    /**
-     * $closure(?PathResult $pathResult) is called, when the path has been found or the pathfinder has a timeout
-     *  => $pathResult is null, when no path was found
-     */
-    public function then(Closure $closure): self {
-        $this->onCompletion = $closure;
-        return $this;
-    }
+					public function __construct(Algorithm $algorithm){
+						$this->algorithm = $algorithm;
+					}
 
-    public function isRunning(): bool{
-        return $this->running;
-    }
+					public function onRun(): void{
+						$this->algorithm->onTick($this->getHandler());
+					}
+				}, 1);
+		} else {
+			$this->onTick(null);
+		}
+		return $this;
+	}
 
-    protected function checkTimout(): bool{
-        $timeout = $this->settings->getTimeout();
-        return $timeout === 0.0 || microtime(true) - $this->startTickTime < $timeout;
-    }
+	public function onTick(?TaskHandler $handler): void{
+		$this->startTickTime = microtime(true);
+		$this->tick();
+		if (--$this->tick <= 0 || !$this->running) {
+			$handler?->cancel();
+			$this->stop();
+		}
+	}
 
-    public function start(): self {
-        if($this->running) return $this;
-        $this->running = true;
-        $this->startTime = microtime(true);
-        $this->tick = $this->settings->getMaxTicks();
+	abstract protected function tick(): void;
 
-        if($this->tick > 0) {
-            Pathfinder::$instance->getScheduler()->scheduleRepeatingTask(
-                new class($this) extends Task {
-                    private Algorithm $algorithm;
+	public function stop(): void{
+		if (!$this->running)
+			return;
+		$this->finish();
+		$this->pathResult?->setTime(round(microtime(true) - $this->startTime, 5));
+		if ($this->onCompletion !== null) {
+			($this->onCompletion)($this->pathResult);
+		}
+		$this->running = false;
+	}
 
-                    public function __construct(Algorithm $algorithm){
-                        $this->algorithm = $algorithm;
-                    }
+	abstract protected function finish(): void;
 
-                    public function onRun(): void{
-                        $this->algorithm->onTick($this->getHandler());
-                    }
-                }, 1
-            );
-        } else {
-            $this->onTick(null);
-        }
-        return $this;
-    }
-
-    public function stop(): void {
-        if(!$this->running) return;
-        $this->finish();
-        $this->pathResult?->setTime(round(microtime(true) - $this->startTime, 5));
-        if($this->onCompletion !== null){
-            ($this->onCompletion)($this->pathResult);
-        }
-        $this->running = false;
-    }
-
-    public function onTick(?TaskHandler $handler): void {
-        $this->startTickTime = microtime(true);
-        $this->tick();
-        if(--$this->tick <= 0 || !$this->running){
-            $handler?->cancel();
-            $this->stop();
-        }
-    }
-
-    abstract protected function finish(): void;
-    abstract protected function tick(): void;
+	protected function checkTimout(): bool{
+		$timeout = $this->settings->getTimeout();
+		return $timeout === 0.0 || microtime(true) - $this->startTickTime < $timeout;
+	}
 }
