+++ src/pathfinder/entity/TestEntity.php
@@ -1,11 +1,8 @@
 <?php
-
 declare(strict_types=1);
-
 namespace pathfinder\entity;
-
 use pathfinder\algorithm\AlgorithmSettings;
-use pathfinder\navigator\Navigator;
+use pathfinder\algorithm\navigator\Navigator;
 use pathfinder\Pathfinder;
 use pocketmine\entity\Location;
 use pocketmine\entity\Villager;
@@ -17,47 +14,42 @@ use Throwable;
 use function array_key_first;
 use function intval;
 
-class TestEntity extends Villager {
-    protected Navigator $navigator;
 
-    public function __construct(Location $location, ?CompoundTag $nbt = null){
-        $this->navigator = new Navigator($this, null, null,
-            (new AlgorithmSettings())
-                ->setTimeout(0.05)
-                ->setMaxTicks(0)
-        );
-        parent::__construct($location, $nbt);
+class TestEntity extends Villager{
+	protected Navigator $navigator;
 
-        $this->setScale(0.5);
-    }
+	public function __construct(Location $location, ?CompoundTag $nbt = null){
+		$this->navigator = new Navigator($this, null, null, (new AlgorithmSettings())->setTimeout(0.05)->setMaxTicks(0));
+		parent::__construct($location, $nbt);
+		$this->setScale(0.5);
+	}
 
-    public function onUpdate(int $currentTick): bool{
-        $target = Server::getInstance()->getOnlinePlayers()[array_key_first(Server::getInstance()->getOnlinePlayers())] ?? null;
-        if($target === null) return parent::onUpdate($currentTick);
-        $position = $target->getPosition();
-        $targetVector3 = $this->navigator->getTargetVector3();
-        if(!$position->world->isInWorld(intval($position->x), intval($position->y), intval($position->z))){
-            return parent::onUpdate($currentTick);
-        }
-
-        if($this->navigator->getTargetVector3() === null || $targetVector3->distanceSquared($position) > 1) {
-            $this->navigator->setTargetVector3($position);
-        }
+	public function onUpdate(int $currentTick): bool{
+		$target = Server::getInstance()->getOnlinePlayers()[array_key_first(Server::getInstance()->getOnlinePlayers())] ?? null;
+		if ($target === null)
+			return parent::onUpdate($currentTick);
+		$position = $target->getPosition();
+		$targetVector3 = $this->navigator->getTargetVector3();
+		if (!$position->world->isInWorld(intval($position->x), intval($position->y), intval($position->z))) {
+			return parent::onUpdate($currentTick);
+		}
+		if ($this->navigator->getTargetVector3() === null || $targetVector3->distanceSquared($position) > 1) {
+			$this->navigator->setTargetVector3($position);
+		}
 		try {
 			$this->navigator->onUpdate();
 		} catch (Throwable $e) {
 			$this->flagForDespawn();
 			Pathfinder::$instance->getLogger()->logException($e);
 		}
-        return parent::onUpdate($currentTick);
-    }
+		return parent::onUpdate($currentTick);
+	}
 
-    public function attack(EntityDamageEvent $source): void{
-        parent::attack($source);
-        $this->setHealth($this->getMaxHealth());
-
-        if($source->getCause() === EntityDamageEvent::CAUSE_FIRE) {
-            $this->flagForDespawn();
-        }
-    }
+	public function attack(EntityDamageEvent $source): void{
+		parent::attack($source);
+		$this->setHealth($this->getMaxHealth());
+		if ($source->getCause() === EntityDamageEvent::CAUSE_FIRE) {
+			$this->flagForDespawn();
+		}
+	}
 }
