--- src/pathfinder/pathpoint/PathPoint.php
@@ -1,10 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace pathfinder\pathpoint;
-
-use pocketmine\math\Vector3;
-
-class PathPoint extends Vector3 {
-}
